Consolidate LinuxUtilities and UnixUtilities.
